@@ -1,9 +1,10 @@
-package net.koofr.koofrexample;
+package net.koofr.android.example;
 
 import android.content.Intent;
 import android.os.AsyncTask;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import android.widget.TextView;
 
@@ -11,8 +12,11 @@ import net.koofr.api.json.JsonException;
 import net.koofr.api.rest.v2.data.Self;
 
 import java.io.IOException;
+import java.lang.ref.WeakReference;
 
 public class UserInfoActivity extends AppCompatActivity {
+    private static final String TAG = UserInfoActivity.class.getName();
+
     TextView nameTextView;
 
     @Override
@@ -20,31 +24,36 @@ public class UserInfoActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_user_info);
 
-        nameTextView = (TextView) findViewById(R.id.user_name);
+        nameTextView = findViewById(R.id.user_name);
 
         nameTextView.setText("Loading...");
 
-        new UserInfoTask().execute();
+        new UserInfoTask(this).execute();
     }
 
     public void logout(View view) {
-        KoofrClient.reset();
+        ((App)getApplication()).getKoofrClient().reset();
 
         Intent intent = new Intent(this, LoginActivity.class);
         startActivity(intent);
     }
 
 
-    private class UserInfoTask extends AsyncTask<Void, Void, Self> {
+    private static class UserInfoTask extends AsyncTask<Void, Void, Self> {
+        WeakReference<UserInfoActivity> activityRef;
+
+        public UserInfoTask(UserInfoActivity activity) {
+            activityRef = new WeakReference<>(activity);
+        }
 
         @Override
         protected Self doInBackground(Void... voids) {
             try {
-                return KoofrClient.getApi().self().get();
+                return ((App)activityRef.get().getApplication()).getKoofrClient().getApi().self().get();
             } catch (JsonException e) {
-                e.printStackTrace();
+                Log.w(TAG, "Failed to fetch user info.", e);
             } catch (IOException e) {
-                e.printStackTrace();
+                Log.w(TAG, "Failed to fetch user info.", e);
             }
 
             return null;
@@ -52,7 +61,7 @@ public class UserInfoActivity extends AppCompatActivity {
 
         @Override
         protected void onPostExecute(Self self) {
-            nameTextView.setText(self.firstName + " " + self.lastName);
+            activityRef.get().nameTextView.setText(self.firstName + " " + self.lastName);
         }
     }
 }
