@@ -1,6 +1,8 @@
-package net.koofr.koofrexample;
+package net.koofr.android.example;
 
-import net.koofr.api.auth.Authenticator;
+import android.content.Context;
+import android.util.Log;
+
 import net.koofr.api.auth.oauth2.OAuth2Authenticator;
 import net.koofr.api.http.Client;
 import net.koofr.api.http.impl.basic.BasicClient;
@@ -11,52 +13,58 @@ import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
 
 public class KoofrClient {
-    private static Api api;
+    private static final String TAG = KoofrClient.class.getName();
 
-    public static String getClientId() {
-        return App.getAppContext().getString(R.string.koofr_client_id);
+    Context ctx;
+    Api api;
+
+    public KoofrClient(Context ctx) {
+        this.ctx = ctx;
+    }
+    public String getClientId() {
+        return ctx.getString(R.string.koofr_client_id);
     }
 
-    public static String getClientSecret() {
-        return App.getAppContext().getString(R.string.koofr_client_secret);
+    public String getClientSecret() {
+        return ctx.getString(R.string.koofr_client_secret);
     }
 
-    public static String getRedirectUrl() {
+    public String getRedirectUrl() {
         return "urn:ietf:wg:oauth:2.0:oob";
     }
 
-    public static String getAuthUrl() {
+    public String getAuthUrl() {
         String encodedRedirectUri = "";
 
         try {
             encodedRedirectUri = URLEncoder.encode(getRedirectUrl(), "UTF-8");
         } catch (UnsupportedEncodingException e) {
-            e.printStackTrace();
+            Log.w(TAG, "URL encoding failed.", e);
         }
 
         return "https://app.koofr.net/oauth2/auth?redirect_uri=" + encodedRedirectUri + "&response_type=code&client_id=" + getClientId() + "&scope=public&platform=android";
     }
 
-    public static String getTokenUrl() {
+    public String getTokenUrl() {
         return "https://app.koofr.net/oauth2/token";
     }
 
-    public static void initialize(String code) throws IOException {
+    public void initialize(String code) throws IOException {
         Client client = new BasicClient();
         OAuth2Authenticator auth = new OAuth2Authenticator(client, getTokenUrl(), getClientId(), getClientSecret(), getRedirectUrl());
         auth.setGrantCode(code);
 
-        // TODO save refresh token
-        // auth.getRefreshToken()
+        // NOTE save refresh token to preferences or use android account framework to store it,
+        // in order to avoid future logins
 
         api = new Api("https://app.koofr.net/api/v2", auth, client);
     }
 
-    public static Api getApi() {
+    public Api getApi() {
         return api;
     }
 
-    public static void reset() {
+    public void reset() {
         api = null;
     }
 }
